refactor(frontend): move businesses list to component
