+++ frontend/src/components/home/HomePage.js
@@ -1,15 +1,10 @@
 import styled, { useTheme } from "styled-components";
-import { Link, useHistory } from "react-router-dom";
 import { useSelector } from "react-redux";
 
-import Card from "../common/Card";
-import CardContent from "../common/CardContent";
 import Container from "../styled/Container";
-import Rating from "../review/Rating";
 import SearchForm from "../search/SearchForm";
-import Typography from "../common/Typography";
 import NestedThemeProvider from "../theme/NestedThemeProvider";
-import { getOverlayAlpha } from "../../utils/theme";
+import BusinessesList from "../business/BusinessesList";
 
 const Hero = styled.div`
   background-image: url(${(props) => props.image});
@@ -31,77 +26,7 @@ const MainContainer = styled(Container).attrs(() => ({ as: "main" }))`
   margin-top: ${(props) => props.theme.spacing.gen(3)};
 `;
 
-const BusinessesList = styled.ul`
-  display: flex;
-  flex-direction: column;
-  gap: ${(props) => props.theme.spacing.gen(3)};
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-  width: 100%;
-`;
-
-const BusinessListItem = styled.li`
-  display: block;
-  width: 100%;
-`;
-
-const CardLink = styled(Link)`
-  text-decoration: none;
-  color: inherit;
-
-  &:hover {
-    text-decoration: underline;
-  }
-`;
-
-const BusinessCard = styled(Card)`
-  background-image: ${(props) =>
-    props.theme.palette.mode === "dark" &&
-    `linear-gradient(rgba(255, 255, 255, ${getOverlayAlpha(
-      1
-    )}), rgba(255, 255, 255, ${getOverlayAlpha(1)}))`};
-  display: flex;
-  gap: ${(props) => props.theme.spacing.gen(2)};
-  cursor: pointer;
-
-  &:hover {
-    background-image: ${(props) =>
-      props.theme.palette.mode === "dark" &&
-      `linear-gradient(rgba(255, 255, 255, ${getOverlayAlpha(
-        2
-      )}), rgba(255, 255, 255, ${getOverlayAlpha(2)}))`};
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  }
-`;
-
-const BusinessImageWrapper = styled.div`
-  height: 220px;
-  width: 220px;
-  padding: ${(props) => props.theme.spacing.gen(2)};
-  padding-right: 0;
-`;
-
-const BusinessImage = styled.img`
-  height: 100%;
-  width: 100%;
-  border-radius: ${(props) => props.theme.borderRadius}px;
-`;
-
-const RatingContainer = styled(Typography).attrs((props) => {
-  return {
-    ...props,
-    as: "div",
-  };
-})`
-  display: flex;
-  align-items: center;
-  gap: ${(props) => props.theme.spacing.gen(1)};
-  color: ${(props) => props.theme.palette.text.secondary};
-`;
-
 const HomePage = () => {
-  const history = useHistory();
   const theme = useTheme();
 
   const businessEntries = useSelector((state) =>
@@ -120,39 +45,7 @@ const HomePage = () => {
         </Hero>
       </NestedThemeProvider>
       <MainContainer>
-        <BusinessesList>
-          {businessEntries.map((business, i) => (
-            <BusinessListItem key={business.id}>
-              <BusinessCard onClick={() => history.push(`/biz/${business.id}`)}>
-                <BusinessImageWrapper>
-                  <BusinessImage
-                    src={business.displayImage}
-                    alt={business.name}
-                  />
-                </BusinessImageWrapper>
-                <CardContent>
-                  <Typography variant="h4" gutterBottom>
-                    {i + 1}.{" "}
-                    <CardLink
-                      to={`/biz/${business.id}`}
-                      onClick={(e) => e.stopPropagation()}
-                    >
-                      {business.name}
-                    </CardLink>
-                  </Typography>
-                  <RatingContainer gutterBottom>
-                    <Rating
-                      rating={business.ratingAverage || 0}
-                      disableButtons
-                      size="small"
-                    />
-                    {business.total || 0}
-                  </RatingContainer>
-                </CardContent>
-              </BusinessCard>
-            </BusinessListItem>
-          ))}
-        </BusinessesList>
+        <BusinessesList businesses={businessEntries} />
       </MainContainer>
     </>
   );
